+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2021.11.22
 
 - Update of dependencies due Vercel indicated bugs
+- TextInput improved to render based on types `text | password | mask |"text-area`, being the `text` the default render
+  - add `antd-mask-input` dependency
 
 ## 2021.11.20